+++ state.go
@@ -1231,7 +1231,7 @@ func (s *State) UserChannelPermissions(userID, channelID string) (apermissions i
 		return
 	}
 
-	return user.MemberPermissions(guild, channel, userID, member.Roles), nil
+	return MemberPermissions(guild, channel, userID, member.Roles), nil
 }
 
 // MessagePermissions returns the permissions of the author of the message
@@ -1255,7 +1255,7 @@ func (s *State) MessagePermissions(message *channel.Message) (apermissions int64
 		return
 	}
 
-	return user.MemberPermissions(guild, channel, message.Author.ID, message.Member.Roles), nil
+	return MemberPermissions(guild, channel, message.Author.ID, message.Member.Roles), nil
 }
 
 // UserColor returns the color of a user in a channel.
