+++ user/member.go
@@ -1,12 +1,9 @@
 package user
 
 import (
-	"github.com/nyttikord/gokord/channel"
-	"github.com/nyttikord/gokord/discord"
-	"github.com/nyttikord/gokord/discord/types"
-	"github.com/nyttikord/gokord/guild"
-
 	"time"
+
+	"github.com/nyttikord/gokord/discord"
 )
 
 // MemberFlags represent flags of a guild member.
@@ -133,71 +130,4 @@ type VoiceState struct {
 	SelfVideo               bool       `json:"self_video"`
 	Suppress                bool       `json:"suppress"`
 	RequestToSpeakTimestamp *time.Time `json:"request_to_speak_timestamp"`
-}
-
-// MemberPermissions calculates the permissions for a user.Member.
-// https://support.discord.com/hc/en-us/articles/206141927-How-is-the-permission-hierarchy-structured-
-func MemberPermissions(guild *guild.Guild, channel *channel.Channel, userID string, roles []string) int64 {
-	if userID == guild.OwnerID {
-		return discord.PermissionAll
-	}
-
-	var perms int64
-	for _, role := range guild.Roles {
-		if role.ID == guild.ID {
-			perms |= role.Permissions
-			break
-		}
-	}
-
-	for _, role := range guild.Roles {
-		for _, roleID := range roles {
-			if role.ID == roleID {
-				perms |= role.Permissions
-				break
-			}
-		}
-	}
-
-	if perms&discord.PermissionAdministrator == discord.PermissionAdministrator {
-		perms |= discord.PermissionAll
-	}
-
-	// Apply @everyone overrides from the channel.
-	for _, overwrite := range channel.PermissionOverwrites {
-		if guild.ID == overwrite.ID {
-			perms &= ^overwrite.Deny
-			perms |= overwrite.Allow
-			break
-		}
-	}
-
-	var denies, allows int64
-	// Member overwrites can override role overrides, so do two passes
-	for _, overwrite := range channel.PermissionOverwrites {
-		for _, roleID := range roles {
-			if overwrite.Type == types.PermissionOverwriteRole && roleID == overwrite.ID {
-				denies |= overwrite.Deny
-				allows |= overwrite.Allow
-				break
-			}
-		}
-	}
-
-	perms &= ^denies
-	perms |= allows
-
-	for _, overwrite := range channel.PermissionOverwrites {
-		if overwrite.Type == types.PermissionOverwriteMember && overwrite.ID == userID {
-			perms &= ^overwrite.Deny
-			perms |= overwrite.Allow
-			break
-		}
-	}
-
-	if perms&discord.PermissionAdministrator == discord.PermissionAdministrator {
-		perms |= discord.PermissionAllChannel
-	}
-
-	return perms
 }
